+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @graphql-hive/nestjs
 
+## 2.0.17
+### Patch Changes
+
+
+
+- [#1637](https://github.com/graphql-hive/gateway/pull/1637) [`aa1cb63`](https://github.com/graphql-hive/gateway/commit/aa1cb635583e41ed0f8ebe7ae4897feedec50a71) Thanks [@dependabot](https://github.com/apps/dependabot)! - dependencies updates:
+  
+  - Updated dependency [`@graphql-mesh/types@^0.104.14` ↗︎](https://www.npmjs.com/package/@graphql-mesh/types/v/0.104.14) (from `^0.104.13`, in `dependencies`)
+- Updated dependencies [[`0e68c01`](https://github.com/graphql-hive/gateway/commit/0e68c01f53fa32ada1575ff7a0b4d31bbe48193b), [`aa1cb63`](https://github.com/graphql-hive/gateway/commit/aa1cb635583e41ed0f8ebe7ae4897feedec50a71)]:
+  - @graphql-hive/gateway@2.1.12
+
 ## 2.0.16
 ### Patch Changes